fix(deploy): restore region env on TUI error states; consolidate barrel exports
Review feedback:
1. TUI preflight error branches called setPhase('error') without calling
   restoreRegionEnv(). Add a useEffect guarded on phase === 'error' so every
   error path restores the env override without threading the call into
   every branch.
2. Export applyTargetRegionToEnv from the aws barrel for consistency with
   withTargetRegion. Update CLI deploy, teardown, and TUI preflight hook
   to import from the barrel instead of the deep path.
